backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
